issue-4690391 refactor models out of summary.r
