+++ app/src/api_models.rs
@@ -1,0 +1,23 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+#[allow(non_snake_case)]
+pub struct BlockchainSummary {
+    pub blockchainLength: u64,
+    pub circulatingSupply: String,
+    pub epoch: u16,
+    pub slot: u16,
+    pub totalCurrency: String,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub enum MyError {
+    NetworkError(String),
+    ParseError(String), // other error variants
+}
+
+impl From<reqwest::Error> for MyError {
+    fn from(err: reqwest::Error) -> Self {
+        MyError::NetworkError(err.to_string())
+    }
+}
